visitor_management_backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
django-filter==23.5
Pillow==10.2.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
reportlab==4.0.8
xlsxwriter==3.1.9
python-dotenv==1.0.0
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0
gunicorn==23.0.0
whitenoise==6.9.0
dj-database-url==2.1.0